Misc int type fixes for -Wconversion, part 2.

*   frequencies (term freq, doc freq)
*   doc counts
*   pointer math results
*   change return type of Sim_Encode_Norm to uint8_t
*   grab bag of other fixes

This closes #49.

+++ core/Lucy/Index/LexiconReader.c
@@ -243,7 +243,7 @@ uint32_t
 DefLexReader_Doc_Freq_IMP(DefaultLexiconReader *self, String *field,
                           Obj *term) {
     TermInfo *tinfo = S_find_tinfo(self, field, term);
-    return tinfo ? TInfo_Get_Doc_Freq(tinfo) : 0;
+    return tinfo ? (uint32_t)TInfo_Get_Doc_Freq(tinfo) : 0;
 }
